Move reusable scripts to the top level
