+++ python/lsst/dax/apdb_migrate/script/migrate_revision.py
@@ -28,7 +28,7 @@ import logging
 from alembic import command, util
 from alembic.script import ScriptDirectory
 
-from ... import config, revision, trees
+from .. import config, revision, trees
 
 _LOG = logging.getLogger(__name__)
